+++ WebPortal/Tenant.Mvc/Core/Models/DiscountModel.cs
@@ -12,7 +12,7 @@ namespace Tenant.Mvc.Core.Models
         public int SeatSectionId { get; set; }
         public int SeatNumber { get; set; }
         public decimal InitialPrice { get; set; }
-        public decimal Discount { get; set; }
+        public int Discount { get; set; }
         public decimal FinalPrice { get; set; }
 
         #endregion
